% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_bool}
\alias{assert_bool}
\title{Assert Boolean Value}
\usage{
assert_bool(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{A character string representing the name of the variable being checked.
This is used in the error message for better debugging.}
}
\value{
None. The function is used for its side effect of throwing an error
  if the input does not meet the criteria.
}
\description{
This function checks if the input is a single, non-missing logical value.
If the input does not meet these criteria, an error is raised.
}
\examples{
assert_bool(TRUE, "my_var")  # Passes validation
assert_bool(FALSE, "my_var") # Passes validation
\dontrun{
assert_bool(NA, "my_var")    # Throws an error
assert_bool(1, "my_var")     # Throws an error
assert_bool(c(TRUE, FALSE), "my_var") # Throws an error
}
}
