% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_inherits}
\alias{assert_inherits}
\title{Assert Object Inherits from a Specific Class}
\usage{
assert_inherits(x, class, name)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A character string specifying the class to check against.}

\item{name}{A character string representing the name of the object (used in the error message).}
}
\value{
This function does not return a value. It is called for its side effect of throwing an error if the check fails.
}
\description{
This function checks whether an object inherits from a specified class. 
If the object does not inherit from the specified class, an error is raised.
}
\examples{
# Example usage:
my_list <- list(a = 1, b = 2)
assert_inherits(my_list, "list", "my_list") # Passes without error

# This will throw an error:
# assert_inherits(my_list, "data.frame", "my_list")

}
