% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Assert Numeric}
\usage{
assert_numeric(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{A character string representing the name of the variable being checked.
This is used in the error message for better debugging.}
}
\value{
None. The function is used for its side effect of throwing an error
  if the input does not meet the criteria.
}
\description{
This function checks if the input is a single, non-missing numeric value.
If the input does not meet these criteria, an error is raised.
}
\examples{
assert_numeric(5, "my_var")  # Passes without error
assert_numeric("a", "my_var")  # Throws an error
assert_numeric(NA, "my_var")  # Throws an error
}
