% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_string}
\alias{assert_string}
\title{Assert that an input is a single non-missing string}
\usage{
assert_string(x, name)
}
\arguments{
\item{x}{The input to check. Expected to be a single string.}

\item{name}{A string representing the name of the input, used in the error message.}
}
\value{
This function does not return a value. It is used for its side effect
of throwing an error if the input is invalid.
}
\description{
This function checks whether the input `x` is a character vector of length 1
and is not `NA`. If the input does not meet these criteria, an error is raised.
}
\examples{
assert_string("example", "input_name") # Passes without error
# assert_string(123, "input_name") # Throws an error
# assert_string(c("a", "b"), "input_name") # Throws an error
# assert_string(NA, "input_name") # Throws an error
}
