% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbid_to_security_id}
\alias{bbid_to_security_id}
\title{Convert Bloomberg Yellow Key to Security ID}
\usage{
bbid_to_security_id(id)
}
\arguments{
\item{yellow_key}{Character. Bloomberg Yellow Key (e.g., "AAPL US Equity").}
}
\value{
Character. Corresponding Security ID.
}
\description{
Convert Bloomberg Yellow Key to Security ID
}
