% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{check_enfusion_connection}
\alias{check_enfusion_connection}
\title{Check connection to Enfusion}
\usage{
check_enfusion_connection()
}
\value{
Bool
}
\description{
Checks connection to Enfusion based on whether a users is logged
 into Enfusion or not. Returns TRUE if logged in, FALSE if not.
}
