% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_rule_compliance}
\alias{check_rule_compliance}
\title{Check Rule Compliance}
\usage{
check_rule_compliance(portfolio_name, update_bbfields = TRUE)
}
\arguments{
\item{portfolio_name}{Character. Name of the portfolio to check.}

\item{update_bbfields}{Logical. Whether to update Bloomberg data before checking rules. Defaults to TRUE.}
}
\description{
Check Rule Compliance
}
