% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_idea_sma.R
\name{create_fake_portfolio}
\alias{create_fake_portfolio}
\title{Create Fake Portfolio}
\usage{
create_fake_portfolio(base_portfolio, long_name, short_name, nav)
}
\arguments{
\item{base_portfolio}{Real portfolio to base fake one on.}

\item{long_name}{Fake portfolio logn name}

\item{short_name}{Fake portfolio short name}

\item{nav}{Target NAV of fake portfolio}
}
\description{
A function to create a fake portfolio based on a specified
 NAV and weights in an existing portfolio
}
