% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio.R
\name{create_portfio}
\alias{create_portfio}
\title{Create Portfolio}
\usage{
create_portfio(id, long_name, short_name, nav, enfusion_url)
}
\arguments{
\item{id}{Id}

\item{long_name}{Long Name}

\item{short_name}{Short Name}

\item{enfusion_url}{enfusion url}

\item{NAV}{NAV}
}
\description{
Create Portfolio
}
