% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio.R
\name{create_portfolio}
\alias{create_portfolio}
\title{Create Portfolio}
\usage{
create_portfolio(long_name, short_name = NULL, nav = NULL, positions = NULL)
}
\arguments{
\item{long_name}{portofolio long name}

\item{short_name}{portfolio short name}

\item{nav}{Portfolio NAV}

\item{positions}{List of positions}
}
\description{
Create an R6 Portfolio object
}
