% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio.R
\name{create_portfolio_from_consolodated_position_report}
\alias{create_portfolio_from_consolodated_position_report}
\title{Create Portfolio from Consoldated Position Report}
\usage{
create_portfolio_from_consolodated_position_report(
  id,
  long_name,
  short_name,
  enfusion_url
)
}
\arguments{
\item{long_name}{Portfolio Long Name}

\item{short_name}{Portfolio Short Name}

\item{enfusion_url}{Enfusion WebServiceURL to Consolodated Positions Report}
}
\value{
portfolio S4 object
}
\description{
Function to create a Portfolio S4 object from an enfusion
 Consolodated Positions Report.
 Specific to the Consolodated Positions Report.
}
