% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio.R
\name{create_portfolio_from_enfusion}
\alias{create_portfolio_from_enfusion}
\title{Create Portfolio from Enfusion}
\usage{
create_portfolio_from_enfusion(long_name, short_name = NULL, enfusion_url)
}
\arguments{
\item{long_name}{Portfolio Long Name}

\item{short_name}{Portfolio Short Name}

\item{enfusion_url}{Enfusion Web URL to}
}
\value{
A \code{Portfolio} object.
}
\description{
Function to create an R6 Portfolio Object
}
