% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_position.R
\name{create_position}
\alias{create_position}
\title{Create Position}
\usage{
create_position(
  portfolio_short_name,
  id,
  qty,
  delta_qty,
  mkt_val,
  delta_val,
  stock_pct_nav,
  delta_pct_nav
)
}
\arguments{
\item{portfolio_short_name}{Portfolio Short Name the security belongs to}

\item{id}{Security Ticker}

\item{qty}{Quantity}

\item{delta_qty}{Delta Quantity}

\item{mkt_val}{Market Value}

\item{delta_val}{Delta Value}

\item{stock_pct_nav}{Stock Percent of NAV}

\item{delta_pct_nav}{Delta Percent of NAV}

\item{desc}{Security Description}
}
\value{
A \code{Position} object.
}
\description{
Create an R6 Position object
}
