% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trade.R
\name{create_proposed_trade_tgt_weight}
\alias{create_proposed_trade_tgt_weight}
\title{Create Proposed Trade Target Weight}
\usage{
create_proposed_trade_tgt_weight(
  portfolio_id = NULL,
  security_id = NULL,
  tgt_weight = NULL,
  swap,
  flow_to_derived = TRUE
)
}
\arguments{
\item{portfolio_id}{[character] The ID of the portfolio. Must be a non-empty string.}

\item{security_id}{[character] The ID of the security. Must be a non-empty string.}

\item{tgt_weight}{[numeric] The target weight for the security in the portfolio. Must be a valid number.}

\item{swap}{[logical] Indicates whether the trade involves a swap.}

\item{flow_to_derived}{[logical] (Default: TRUE) Determines if the flow should propagate to derived positions.}
}
\value{
[invisible(data.frame)] A data frame containing the proposed trade details.
}
\description{
This function calculates the proposed trade quantity based on the target weight 
for a given security in a portfolio and creates a proposed trade data frame.
}
\details{
The function calculates the target quantity of the security based on the portfolio's 
net asset value (NAV) and the security's price. It then computes the trade quantity 
required to achieve the target weight by comparing it with the current quantity. 
Finally, it creates a proposed trade data frame using the calculated trade quantity.
}
\examples{
\dontrun{
create_proposed_trade_tgt_weight(
  portfolio_id = "portfolio123",
  security_id = "security456",
  tgt_weight = 0.05,
  swap = FALSE,
  flow_to_derived = TRUE
)
}

}
\seealso{
\code{\link{create_proposed_trade_qty}}
}
