% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scaled_position.R
\name{create_scaled_position}
\alias{create_scaled_position}
\title{Create a Scaled Position (R6 Object)}
\usage{
create_scaled_position(position, scaling_factor)
}
\arguments{
\item{position}{position to be scaled}

\item{scaling_factor}{scaling_factor}
}
\description{
Create a Scaled Position (R6 Object)
}
