% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_security.R
\name{create_security}
\alias{create_security}
\title{Create Security}
\usage{
create_security(
  id = NULL,
  description = NULL,
  instrument_type = NULL,
  occ_code = NULL,
  ...
)
}
\arguments{
\item{id}{Security Ticker}

\item{description}{Security Description}

\item{instrument_type}{security instrument type}

\item{occ_code}{security occ_code}

\item{...}{Additional data}
}
\value{
A \code{Security} object.
}
\description{
Function to create an R6 Security Object
}
