% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sma.R
\name{create_sma}
\alias{create_sma}
\title{Create SMA}
\usage{
create_sma(
  long_name,
  short_name,
  base_portfolio = NULL,
  enfusion_url = NULL,
  nav = NULL
)
}
\arguments{
\item{long_name}{Long name of SMA object.}

\item{short_name}{Short identifying name of SMA object.}

\item{base_portfolio}{Short name of portfolio the SMA is
intended to mimick. This portfolio must already be created for
the SMA to be created.}

\item{enfusion_url}{Optional Web URL to Enfusion
Consolidated Position Listing Report. If supplied, current
positions in the SMA will be created based on the holdings
in the enfusion report. If not supplied, NAV must be supplied.}

\item{nav}{Optional NAV of SMA. If enfusion_url is supplied, arguement will
be ignored.}
}
\value{
A \code{SMA} object.
}
\description{
Fucntion to create an R6 SMA Object.
}
