% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_enfusion.R
\name{create_sma_from_enfusion}
\alias{create_sma_from_enfusion}
\title{Create Portfolio from Enfusion}
\usage{
create_sma_from_enfusion(
  long_name,
  short_name = NULL,
  base_portfolio,
  enfusion_url
)
}
\arguments{
\item{long_name}{Portfolio Long Name}

\item{short_name}{Portfolio Short Name}

\item{enfusion_url}{Enfusion Web URL to}
}
\value{
A \code{Portfolio} object.
}
\description{
Function to create an R6 Portfolio Object
}
