% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sma_rule.R
\name{create_sma_rule}
\alias{create_sma_rule}
\title{Create SMA Rule}
\usage{
create_sma_rule(name, scope = c("agg", "ind"), desc, func)
}
\arguments{
\item{name}{Rule name}

\item{scope}{Rule scope}

\item{desc}{Rule description}

\item{func}{Rule function}
}
\description{
Create an SMA Rule
}
