% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_smarule.R
\name{create_smarule}
\alias{create_smarule}
\title{Create SMA Rule}
\usage{
create_smarule(
  sma_name = NULL,
  rule_name = NULL,
  scope = NULL,
  definition = NULL,
  threshold = NULL
)
}
\arguments{
\item{sma_name}{Name of SMA the rule applies to}

\item{rule_name}{Name of the SMA rule}

\item{scope}{Scope of the rule (position)}

\item{definition}{Function to define the rule}

\item{threshold}{Function to define the threshold}
}
\description{
Function to create a SMA Rule object
}
