% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trade.R
\name{create_trade}
\alias{create_trade}
\title{Create a new object of class \code{Trade}}
\usage{
create_trade(security_id, portfolio_id)
}
\arguments{
\item{security_id}{The security ID of the trade.}

\item{portfolio_id}{The portfolio ID of the trade.}
}
\description{
Function to create a new object of class \code{Trade} from a list of arguments.
}
