% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trade.R
\name{create_trade_qty}
\alias{create_trade_qty}
\title{Create a Trade with Specified Quantity}
\usage{
create_trade_qty(
  portfolio_id = NULL,
  security_id = NULL,
  trade_qty = NULL,
  swap
)
}
\arguments{
\item{portfolio_id}{A string representing the ID of the portfolio. Default is `NULL`.}

\item{security_id}{A string representing the ID of the security. Default is `NULL`.}

\item{trade_qty}{A numeric value specifying the quantity of the trade. Default is `NULL`.}

\item{swap}{A boolean indicating whether the trade involves a swap. If not provided, it is inferred from the portfolio's position.}
}
\value{
An invisible trade object created with the specified parameters.
}
\description{
This function creates a trade for a given portfolio and security with a specified quantity. 
It also handles swap logic and updates derived portfolios if applicable.
}
\details{
The function validates the input parameters and retrieves the security and portfolio objects. 
If the `swap` parameter is not explicitly provided, it attempts to infer it from the portfolio's position. 
After creating the trade, the function checks for any derived portfolios and ensures they mimic the base portfolio's 
position for the given security.
}
\examples{
\dontrun{
create_trade_qty(
  portfolio_id = "portfolio123",
  security_id = "security456",
  trade_qty = 100,
  swap = TRUE
)
}

}
