% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trade.R
\name{create_trade_tgt_weight}
\alias{create_trade_tgt_weight}
\title{Create a Trade Based on Target Weight}
\usage{
create_trade_tgt_weight(
  portfolio_id = NULL,
  security_id = NULL,
  tgt_weight = NULL,
  swap
)
}
\arguments{
\item{portfolio_id}{[character] The ID of the portfolio. Must be a non-empty string.}

\item{security_id}{[character] The ID of the security. Must be a non-empty string.}

\item{tgt_weight}{[numeric] The target weight for the security in the portfolio. Must be a valid number.}

\item{swap}{[logical] Indicates whether the trade involves a swap. If not provided, it will be inferred from the portfolio's current position.}
}
\value{
Creates a trade and returns the result of the `create_trade_qty` function.
}
\description{
This function creates a trade for a given portfolio and security based on a target weight.
}
\details{
The function calculates the target quantity of the security based on the portfolio's net asset value (NAV) and the security's price.
It then determines the difference between the target quantity and the current quantity to compute the trade quantity.
Finally, it creates a trade with the calculated trade quantity.
}
\examples{
\dontrun{
create_trade_tgt_weight(
  portfolio_id = "portfolio123",
  security_id = "security456",
  tgt_weight = 0.05,
  swap = TRUE
)
}

}
\seealso{
\code{\link{create_trade_qty}}, \code{\link{.portfolio}}, \code{\link{.security}}
}
