% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from Postgres Database}
\usage{
db_disconnect()
}
\value{
\code{TRUE} invisibly.
}
\description{
Disconnects the cached connection and clears it from the package
environment.
}
