% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{.bulk_holding_positions}
\alias{.bulk_holding_positions}
\title{Internal helper: create many holdings from Enfusion report}
\usage{
.bulk_holding_positions(enfusion_report, portfolio_short_name)
}
\arguments{
\item{enfusion_report}{Data frame of Enfusion report rows}

\item{short_name}{Character portfolio short name}
}
\description{
Internal helper: create many holdings from Enfusion report
}
