% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{.bulk_trade_positions}
\alias{.bulk_trade_positions}
\title{Internal helper: create many positions from Enfusion report}
\usage{
.bulk_trade_positions(trade_url, portfolio)
}
\arguments{
\item{enfusion_report}{Data frame of Enfusion report rows}

\item{short_name}{Character portfolio short name}
}
\description{
Internal helper: create many positions from Enfusion report
}
