% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.check_portfolio_db_id}
\alias{.check_portfolio_db_id}
\title{Check if Portfolio Exists in Database}
\usage{
.check_portfolio_db_id(portfolio_db_id, con = get_db_connection())
}
\arguments{
\item{portfolio_db_id}{The portfolio_id to check for existence in the
database.}

\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns TRUE if the portfolio exists, FALSE otherwise.
}
\description{
Checks if a portfolio with the given portfolio_db_id exists in
the database.
}
