% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.fetch_holdings_from_db}
\alias{.fetch_holdings_from_db}
\title{Load Holdings from DB}
\usage{
.fetch_holdings_from_db(portfolio_id, con = get_db_connection())
}
\arguments{
\item{portfolio_id}{Portfolio Id}

\item{con}{connection}
}
\description{
Load Holdings from DB
}
