% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.fetch_ordcon_from_db}
\alias{.fetch_ordcon_from_db}
\title{Load Order Constructor data}
\usage{
.fetch_ordcon_from_db(portfolio_id, con = get_db_connection())
}
\arguments{
\item{portfolio_id}{Portfolio Id}

\item{con}{connection}
}
\description{
Load Order Constructor data
}
