% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.fetch_portfolio_id_by_short_name}
\alias{.fetch_portfolio_id_by_short_name}
\title{Fetch Portfolio ID by Short Name}
\usage{
.fetch_portfolio_id_by_short_name(short_name, con = get_db_connection())
}
\arguments{
\item{short_name}{The short name of the portfolio to look up.}

\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns the portfolio_id as an integer if found, otherwise throws an
error.
}
\description{
Retrieves the portfolio_id from the database for a given
portfolio short name.
}
