% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.fetch_portfolio_short_name_by_id}
\alias{.fetch_portfolio_short_name_by_id}
\title{Fetch Portfolio Short Name by ID}
\usage{
.fetch_portfolio_short_name_by_id(portfolio_id, con = get_db_connection())
}
\arguments{
\item{portfolio_id}{The portfolio_id to look up.}

\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns the short name of the portfolio as a character string if 
found, otherwise throws an error.
}
\description{
Retrieves the short name of a portfolio from the database for a
given portfolio_id.
}
