% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.fetch_sma_rules_from_db}
\alias{.fetch_sma_rules_from_db}
\title{Load SMA Rules from DB}
\usage{
.fetch_sma_rules_from_db(portfolio_id, con = get_db_connection())
}
\arguments{
\item{portfolio_id}{Portfolio Id (integer or short name string)}

\item{con}{connection}
}
\description{
Load SMA Rules from DB
}
