% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.load_holdings_from_db}
\alias{.load_holdings_from_db}
\title{Load Holdings from DB into a Portfolio's Registry Entry}
\usage{
.load_holdings_from_db(con, portfolio_short_name, portfolio_db_id, date)
}
\arguments{
\item{con}{DBI connection}

\item{portfolio_short_name}{Character portfolio short name}

\item{portfolio_db_id}{Integer portfolio_id in the \code{portfolios} table}

\item{date}{Date}
}
\description{
Load Holdings from DB into a Portfolio's Registry Entry
}
\keyword{internal}
