% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.load_rules_from_db}
\alias{.load_rules_from_db}
\title{Load Rules for an SMA from Database and Attach to the SMA Object}
\usage{
.load_rules_from_db(con, portfolio_short_name, portfolio_db_id)
}
\arguments{
\item{con}{DBI connection}

\item{portfolio_short_name}{Character short name of the SMA}

\item{portfolio_db_id}{Integer portfolio_id}
}
\description{
Load Rules for an SMA from Database and Attach to the SMA Object
}
\keyword{internal}
