% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.load_sma_rules_from_db}
\alias{.load_sma_rules_from_db}
\title{Load SMA Rules from DB}
\usage{
.load_sma_rules_from_db(portfolio_id, con = get_db_connection())
}
\arguments{
\item{con}{connection}

\item{porfolios_id}{Portfolio_id}
}
\description{
Load SMA Rules from DB
}
