% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{.make_positions}
\alias{.make_positions}
\title{Internal helper: create positions in parallel}
\usage{
.make_positions(enfusion_report, short_name, position_fn)
}
\arguments{
\item{enfusion_report}{Data frame of Enfusion report rows}

\item{short_name}{Character portfolio short name}

\item{position_fn}{Function(x, portfolio_short_name) that returns a Position}
}
\value{
List of Position objects
}
\description{
Internal helper: create positions in parallel
}
