% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{.position}
\alias{.position}
\title{Create or Retrieve a Position Object}
\usage{
.position(portfolio_name, sec_id, create = FALSE, assign_to_portfolio = FALSE)
}
\arguments{
\item{portfolio_name}{A string specifying the name of the portfolio.
The portfolio must already exist.}

\item{sec_id}{A string representing the Bloomberg identifier (BBID) of the
security. This will be converted to lowercase.}

\item{create}{A logical value indicating whether to create the position}

\item{assign_to_portfolio}{A logical value indicating whether to add the
position to the portfolio. Defaults to `TRUE`.}
}
\value{
An object of class `Position` representing the created or updated
 position.
}
\description{
This function creates or updates a position in a specified portfolio.
 It ensures that the input parameters are valid and initializes a new
 position object.
}
\details{
The function validates the input parameters using assertion checks.
 If the security does not already exist, it will be created. The position is
 then initialized using the `Position$new` method.
}
\examples{
# Create a position with 100 shares of a security
.position("MyPortfolio", "AAPL US Equity")

}
\seealso{
\code{\link{Position}} for the Position class.
}
