% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.update_db_holdings_data}
\alias{.update_db_holdings_data}
\title{Update Holdings Data in Database}
\usage{
.update_db_holdings_data(
  holdings_url =
    "https://webservices.enfusionsystems.com/mobile/rest/reportservice/exportReport?name=shared\%2FTaylor\%2FSMA_Mgr_Reports\%2FPosition.ppr",
  con = get_db_connection()
)
}
\arguments{
\item{holdings_url}{URL to fetch the Enfusion holdings report. Default is set
to the shared Taylor SMA Manager Reports Position report.}

\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns TRUE invisibly if the update is successful, or FALSE if the
connection is NULL.
}
\description{
Updates the holdings data in the database by fetching the latest
Enfusion report and writing it to the database. This function requires a
valid database connection.
}
