% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.update_db_nav_data}
\alias{.update_db_nav_data}
\title{Update NAV Data in Database}
\usage{
.update_db_nav_data(
  nav_url =
    "https://webservices.enfusionsystems.com/mobile/rest/reportservice/exportReport?name=shared\%2FTaylor\%2FSMA_Mgr_Reports\%2FNAV.ppr",
  con = get_db_connection()
)
}
\arguments{
\item{nav_url}{URL to fetch the Enfusion NAV report. Default is set
to the shared Taylor SMA Manager Reports NAV report.}

\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns TRUE invisibly if the update is successful, or FALSE if the
connection is NULL.
}
\description{
Updates the NAV data in the database by fetching the latest
Enfusion report and writing it to the database. This function requires a
valid database connection.
}
