% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{.update_db_trade_data}
\alias{.update_db_trade_data}
\title{Update Trade Data in Database}
\usage{
.update_db_trade_data(
  trades_url =
    "https://webservices.enfusionsystems.com/mobile/rest/reportservice/exportReport?name=shared\%2FTaylor\%2FSMA_Mgr_Reports\%2FSMA_Trade_Detail.trb",
  con = get_db_connection()
)
}
\arguments{
\item{trades_url}{URL to fetch the Enfusion trades report. Default is set
to the shared Taylor SMA Manager Reports Trade Detail report.
#' @param con A valid DBI connection object to the Postgres database.}
}
\value{
Returns TRUE invisibly if the update is successful, or FALSE if the
connection is NULL.
}
\description{
Updates the trade data in the database by fetching the latest
Enfusion report and writing it to the database. This function requires a
valid database connection.
}
