% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{get_portfolio}
\alias{get_portfolio}
\title{Get Portfolio}
\usage{
get_portfolio(short_name = NULL)
}
\arguments{
\item{short_name}{Portfolio Short Name}
}
\description{
Function to get R6 Portfolio from environment
}
