% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_security.R
\name{get_security}
\alias{get_security}
\title{Get Security (R6 Object)}
\usage{
get_security(ticker = NULL)
}
\arguments{
\item{ticker}{Security Ticker}
}
\description{
Function to get R6 Security Object from envir
}
