% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-securitydataprovider.R
\name{get_security_data_provider}
\alias{get_security_data_provider}
\title{Get the Active Security Data Provider}
\usage{
get_security_data_provider()
}
\value{
A \code{\link{SecurityDataProvider}} object.
}
\description{
Returns the provider set by
\code{\link{set_security_data_provider}}. If none has been set, a
\code{\link{BloombergDataProvider}} is created, cached, and returned.
}
