% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracking_smas.R
\name{get_tracking_smas}
\alias{get_tracking_smas}
\title{Get Tracking SMAs for a Portfolio}
\usage{
get_tracking_smas(portfolio = NULL)
}
\arguments{
\item{portfolio}{Portfolio object. The portfolio for which to get the tracking SMAs.}
}
\value{
A list of tracking SMAs for the given portfolio. If no tracking SMAs are found, NULL is returned.
}
\description{
Get tracking SMAs for a given portfolio.
}
