% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trade.R
\name{get_trade}
\alias{get_trade}
\title{Get Trade}
\usage{
get_trade(security_id = NULL, swap = FALSE)
}
\arguments{
\item{security_id}{Security ID}

\item{swap}{Is trade on swap? Default is FALSE}
}
\description{
Get trade from trade environment
}
