% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_history.R
\name{get_yahoo_history}
\alias{get_yahoo_history}
\title{Get Yahoo Finance Historical Data}
\usage{
get_yahoo_history(symbol = NULL)
}
\arguments{
\item{symbol}{Ticker Symbol Used on Yahoo Finance}
}
\description{
Get Yahoo Finance Historical Data
}
