% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null/empty coalescing
Returns y if x is NULL, length 0, or NA; otherwise returns x.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{An object to check.}

\item{y}{A fallback value to return if x is NULL, length 0, or NA.}
}
\description{
Null/empty coalescing
Returns y if x is NULL, length 0, or NA; otherwise returns x.
}
