% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\docType{class}
\name{portfolio-class}
\alias{portfolio-class}
\title{Portfolio (S4 Object)}
\description{
S4 Class representing a portfolio object
}
\section{Slots}{

\describe{
\item{\code{id}}{Account Id}

\item{\code{long_name}}{Portfolio Long Name}

\item{\code{short_name}}{Portfolio Short Name}

\item{\code{enfusion_url}}{Enfusion Web URL to Consolidated Position Listing Report}

\item{\code{positions}}{List of Positions}
}}

