% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{read_yml_portfolio}
\alias{read_yml_portfolio}
\title{Read Portfolio Definition from YAML File}
\usage{
read_yml_portfolio(file)
}
\arguments{
\item{file}{Path to YAML file}
}
\value{
Portfolio or SMA object
}
\description{
Read Portfolio Definition from YAML File
}
