% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_sma.R
\name{rebalance_sma}
\alias{rebalance_sma}
\title{Rebalance SMA}
\usage{
rebalance_sma(sma = NULL)
}
\arguments{
\item{sma}{SMA Portfolio object}
}
\value{
Null
}
\description{
Function to rebalance an SMA portfolio to match the base portfolio
}
