% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_sma.R
\name{rebalance_sma_position}
\alias{rebalance_sma_position}
\title{Rebalance SMA Position}
\usage{
rebalance_sma_position(sma, base_position_id)
}
\arguments{
\item{sma}{SMA Portfolio object}

\item{position_id}{Position ID to rebalance}
}
\value{
Null
}
\description{
Function to rebalance a position in an SMA portfolio to match
 the target portfolio
}
