% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_all_portfolio_trades.R
\name{remove_all_portfolio_trades}
\alias{remove_all_portfolio_trades}
\title{Remove all trades for a portfolio}
\usage{
remove_all_portfolio_trades(portfolio_id = NULL)
}
\arguments{
\item{portfolio_id}{The portfolio ID of the trade.}
}
\description{
Remove all trades for a portfolio
}
