% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\docType{class}
\name{security-class}
\alias{security-class}
\title{Security (S4 Object)}
\description{
An S4 Class to represent a security
}
\section{Slots}{

\describe{
\item{\code{ticker}}{Security Ticker}

\item{\code{yahoo_match}}{Boolean if found on yahoo finance}

\item{\code{long_name}}{Security Long Name}

\item{\code{short_name}}{Security Short Name}

\item{\code{adv}}{Average Daily Volume}

\item{\code{dates}}{Dates corresponding to data}

\item{\code{open}}{Open Prices}

\item{\code{close}}{Close Prices}

\item{\code{low}}{Low Prices}

\item{\code{high}}{High Prices}

\item{\code{volume}}{Trading Volume}

\item{\code{adj_close}}{Adjusted Close Price}
}}

