% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\docType{class}
\name{SecurityR6}
\alias{SecurityR6}
\title{Security (R6 Object)}
\description{
A security where data comes from yahoo finance
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ticker}}{Security Ticker}

\item{\code{yahoo_match}}{Boolean if ticker was found on yahoo finance}

\item{\code{long_name}}{Security Long Name}

\item{\code{short_name}}{Security Short Name}

\item{\code{adv}}{Average Daily Volume}

\item{\code{dates}}{Dates corresponding to data}

\item{\code{open}}{Open Prices}

\item{\code{close}}{Close Prices}

\item{\code{low}}{Low Prices}

\item{\code{high}}{High Prices}

\item{\code{volume}}{Trading Volume}

\item{\code{adj_close}}{Adjusted Close Price}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecurityR6-new}{\code{SecurityR6$new()}}
\item \href{#method-SecurityR6-clone}{\code{SecurityR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityR6-new}{}}}
\subsection{Method \code{new()}}{
Create New SecurityR6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityR6$new(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Security Ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SecurityR6` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecurityR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecurityR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecurityR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
