% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\docType{class}
\name{sma-class}
\alias{sma-class}
\title{SMA (S4 Object)}
\description{
S4 Class representing a separetly managed account
}
\section{Slots}{

\describe{
\item{\code{id}}{Account Id}

\item{\code{long_name}}{SMA Long Name}

\item{\code{short_name}}{SMA Short Name}

\item{\code{positions}}{List of Positions}
}}

