% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_rule.R
\docType{class}
\name{sma_rule-class}
\alias{sma_rule-class}
\title{SMA Rule (S4 Object)}
\description{
Parent Class for SMA Rules
}
\section{Slots}{

\describe{
\item{\code{rule_id}}{Rule ID}

\item{\code{rule_name}}{Rule Name}

\item{\code{rule_scope}}{Rule Scope (Agg/SMA)}

\item{\code{rule_desc}}{Rule Type}
}}

