% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_rule.R
\docType{class}
\name{sma_rule_port-class}
\alias{sma_rule_port-class}
\title{SMA Position Rule (S4 Object)}
\description{
SMA Rule about positions in SMA
}
\section{Slots}{

\describe{
\item{\code{position_func}}{Function to be applied to functions}
}}

