% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trade.R
\name{Trade}
\alias{Trade}
\title{Trade (R6 Object)}
\description{
R6 Class representing a trade object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Trade-new}{\code{Trade$new()}}
\item \href{#method-Trade-get_id}{\code{Trade$get_id()}}
\item \href{#method-Trade-get_portfolio}{\code{Trade$get_portfolio()}}
\item \href{#method-Trade-get_position}{\code{Trade$get_position()}}
\item \href{#method-Trade-add_trade}{\code{Trade$add_trade()}}
\item \href{#method-Trade-allocate_trade}{\code{Trade$allocate_trade()}}
\item \href{#method-Trade-to_df}{\code{Trade$to_df()}}
\item \href{#method-Trade-clone}{\code{Trade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-new"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-new}{}}}
\subsection{Method \code{new()}}{
Create New Trade R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$new(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get trade id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$get_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-get_portfolio"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-get_portfolio}{}}}
\subsection{Method \code{get_portfolio()}}{
Get trade portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$get_portfolio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-get_position}{}}}
\subsection{Method \code{get_position()}}{
Get trade position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$get_position()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-add_trade"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-add_trade}{}}}
\subsection{Method \code{add_trade()}}{
Add Trade to Trade Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$add_trade(portfolio_short_name = NULL, quantity = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{portfolio_short_name}}{Portfolio Short Name}

\item{\code{quantity}}{Quantity of Trade}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Null
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-allocate_trade"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-allocate_trade}{}}}
\subsection{Method \code{allocate_trade()}}{
Calculate Trade Allocation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$allocate_trade()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Null
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-to_df"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-to_df}{}}}
\subsection{Method \code{to_df()}}{
Output to data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$to_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame of trade allocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trade-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Trade-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
