% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tradeconstructor.R
\name{TradeConstructor}
\alias{TradeConstructor}
\title{Portfolio Constructor}
\description{
Portfolio Constructor

Portfolio Constructor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TradeConstructor-get_security_position_limits}{\code{TradeConstructor$get_security_position_limits()}}
\item \href{#method-TradeConstructor-get_swap_flag_position_rules}{\code{TradeConstructor$get_swap_flag_position_rules()}}
\item \href{#method-TradeConstructor-calc_trade_qty}{\code{TradeConstructor$calc_trade_qty()}}
\item \href{#method-TradeConstructor-clone}{\code{TradeConstructor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TradeConstructor-get_security_position_limits"></a>}}
\if{latex}{\out{\hypertarget{method-TradeConstructor-get_security_position_limits}{}}}
\subsection{Method \code{get_security_position_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TradeConstructor$get_security_position_limits(portfolio, security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{portfolio}}{An object of class Portfolio}

\item{\code{security_id}}{A character vector of security IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TradeConstructor-get_swap_flag_position_rules"></a>}}
\if{latex}{\out{\hypertarget{method-TradeConstructor-get_swap_flag_position_rules}{}}}
\subsection{Method \code{get_swap_flag_position_rules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TradeConstructor$get_swap_flag_position_rules(portfolio, security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{portfolio}}{An object of class Portfolio}

\item{\code{security_id}}{A character vector of security IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TradeConstructor-calc_trade_qty"></a>}}
\if{latex}{\out{\hypertarget{method-TradeConstructor-calc_trade_qty}{}}}
\subsection{Method \code{calc_trade_qty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TradeConstructor$calc_trade_qty(
  portfolio,
  target_sec_id = NULL,
  target_trade_qty = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{portfolio}}{An object of class Portfolio}

\item{\code{target_sec_id}}{A character vector of target security IDs}

\item{\code{target_trade_qty}}{A numeric vector of target trade quantities}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TradeConstructor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TradeConstructor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TradeConstructor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
