% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{update_db_data}
\alias{update_db_data}
\title{Update Database with Latest Enfusion Data}
\usage{
update_db_data(con = get_db_connection())
}
\arguments{
\item{con}{A valid DBI connection object to the Postgres database.}
}
\value{
Returns TRUE invisibly if the update is successful, or FALSE if the
connection is NULL. If the connection is NULL, the function will not attempt
to update the database and will return FALSE immediately.
}
\description{
Updates the database with the latest holdings and trade data
from Enfusion. This function requires a valid database connection.
}
