% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{create_portfolio_from_enfusion}
\alias{create_portfolio_from_enfusion}
\title{Create Portfolio from Enfusion}
\usage{
create_portfolio_from_enfusion(long_name, short_name, holdings_url)
}
\arguments{
\item{long_name}{Character long name}

\item{short_name}{Character short name}

\item{holdings_url}{URL to fetch Enfusion report}
}
\value{
Portfolio R6 object
}
\description{
Create Portfolio from Enfusion
}
