% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_position.R
\name{create_position}
\alias{create_position}
\title{Create Position}
\usage{
create_position(
  ticker,
  desc,
  stock_qty,
  delta_qty,
  total_qty,
  mkt_val,
  delta_val,
  stock_pct_nav,
  delta_pct_nav
)
}
\arguments{
\item{ticker}{Ticker}

\item{desc}{Description}

\item{stock_qty}{Stock Quantity}

\item{delta_qty}{Delta Quantity}

\item{total_qty}{Total Quantity}

\item{mkt_val}{Market Value}

\item{delta_val}{Delta Value}

\item{stock_pct_nav}{Stock Percent of NAV}

\item{delta_pct_nav}{Delta Percent of NAV}
}
\description{
Create Position
}
