% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{create_position_from_enfusion}
\alias{create_position_from_enfusion}
\title{Create a Position from Enfusion Data}
\usage{
create_position_from_enfusion(x, portfolio_short_name)
}
\arguments{
\item{x}{A list or single-row data.frame of enfusion data}

\item{portfolio_short_name}{Character short name of the portfolio}
}
\value{
Position R6 object
}
\description{
Create a Position from Enfusion Data
}
