% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_security.R
\name{create_security}
\alias{create_security}
\title{Create Security}
\usage{
create_security(ticker, adv_days = 30)
}
\arguments{
\item{ticker}{Security Ticker}

\item{adv_days}{ADV Days}
}
\description{
Create Security
}
