% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_security.R
\name{create_security}
\alias{create_security}
\title{Create Security}
\usage{
create_security(ticker = NULL)
}
\arguments{
\item{ticker}{Security Ticker}
}
\value{
A \code{Security} object.
}
\description{
Function to create an R6 Security Object
}
