% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfusion_loading.R
\name{create_sma_from_enfusion}
\alias{create_sma_from_enfusion}
\title{Create SMA from Enfusion}
\usage{
create_sma_from_enfusion(
  long_name,
  short_name,
  base_portfolio,
  holdings_url,
  trade_url
)
}
\arguments{
\item{long_name}{Character long name}

\item{short_name}{Character short name}

\item{base_portfolio}{Portfolio object}

\item{enfusion_url}{URL to fetch Enfusion report}
}
\value{
SMA R6 object
}
\description{
Create SMA from Enfusion
}
