% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_smarule.R
\name{create_smarule}
\alias{create_smarule}
\title{Create SMA Rule}
\usage{
create_smarule(
  sma_name = NULL,
  rule_name = NULL,
  scope = NULL,
  definition = NULL,
  max_threshold = Inf,
  min_threshold = -Inf,
  swap_only = FALSE
)
}
\arguments{
\item{sma_name}{Name of SMA the rule applies to}

\item{rule_name}{Name of the SMA rule}

\item{scope}{Scope of the rule (position)}

\item{definition}{Function to define the rule}

\item{max_threshold}{Max weight of the rule}

\item{min_threshold}{Min weight of the rule}

\item{swap_only}{If TRUE, position can only be on swap}
}
\description{
Function to create a SMA Rule object
}
