% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{.position}
\alias{.position}
\title{Create or Update a Position in a Portfolio}
\usage{
.position(portfolio_name, bbid, qty = 0, swap = FALSE)
}
\arguments{
\item{portfolio_name}{A string specifying the name of the portfolio.
The portfolio must already exist.}

\item{bbid}{A string representing the Bloomberg identifier (BBID) of the
security. This will be converted to lowercase.}

\item{qty}{A numeric value indicating the quantity of the position.
Defaults to 0.}

\item{swap}{A logical value (TRUE or FALSE) indicating whether the position
is a swap. Defaults to FALSE.}
}
\value{
An object of class `Position` representing the created or updated
 position.
}
\description{
This function creates or updates a position in a specified portfolio.
 It ensures that the input parameters are valid and initializes a new
 position object.
}
\details{
The function validates the input parameters using assertion checks.
 If the security does not already exist, it will be created. The position is
 then initialized using the `Position$new` method.
}
\examples{
# Create a position with 100 shares of a security
position("MyPortfolio", "AAPL US Equity", qty = 100)

# Create a swap position
position("MyPortfolio", "AAPL US Equity", qty = 50, swap = TRUE)

}
\seealso{
\code{\link{Position}} for the Position class.
}
