% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{.security}
\alias{.security}
\title{Retrieve or Create a Security Object}
\usage{
.security(bbid, create = TRUE)
}
\arguments{
\item{bbid}{A string representing the Bloomberg ID of the security. Must be a non-empty string.}

\item{create}{A logical value indicating whether to create a new security object if it does not
already exist in the registry. Defaults to `TRUE`.}
}
\value{
If the security exists in the registry, it returns the corresponding security object.
  If the security does not exist and `create` is `FALSE`, it returns `NULL`. If `create` is `TRUE`,
  it creates a new security object, adds it to the registry, and returns it.
}
\description{
This function retrieves a security object from the securities registry by its Bloomberg ID (bbid).
If the security does not exist in the registry and `create` is set to `TRUE`, it attempts to create
a new security object using Bloomberg data.
}
\details{
The function first validates the `bbid` parameter to ensure it is a valid string. It then
  checks if the security exists in the `registries$securities` environment. If the security does not
  exist and `create` is `TRUE`, it queries Bloomberg for the security's data using the `Rblpapi::bpd`
  function. If the Bloomberg data indicates the security is not found, an error is raised. Otherwise,
  a new `Security` object is created and added to the registry.
}
\examples{
# Retrieve an existing security
security("AAPL US Equity")

# Create a new security if it does not exist
security("MSFT US Equity", create = TRUE)

# Attempt to retrieve a security without creating it
security("GOOG US Equity", create = FALSE)

}
\seealso{
\code{\link{Security}} for the Security class.
}
