% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_create_security.R
\name{get_or_create_security}
\alias{get_or_create_security}
\title{get or Create Security}
\usage{
get_or_create_security(id = NULL, ...)
}
\value{
A \code{Security} object.
}
\description{
Function to get or create R6 Security Object.
If security already exists, return object.
If securtiy does not already exist, create object
}
