% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tracking_portfolios}
\alias{get_tracking_portfolios}
\title{Get Tracking Portfolios}
\usage{
get_tracking_portfolios(base_portfolio)
}
\arguments{
\item{base_portfolio}{Character short name or Portfolio object}
}
\value{
list of SMA portfolios
}
\description{
Get all SMAs that have a specified portfolio as any part of
  their base (including blended bases).
}
