% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracking_smas.R
\name{get_tracking_smas}
\alias{get_tracking_smas}
\title{Get Tracking SMAs}
\usage{
get_tracking_smas(portfolio)
}
\arguments{
\item{portfolio}{An object of class `Portfolio`. The portfolio for which 
tracking SMAs are to be retrieved.}
}
\value{
A list of SMA objects that are tracking the given portfolio. If no 
  tracking SMAs are found, the function returns `NULL`.
}
\description{
This function retrieves all SMA (Separately Managed Account) objects that are 
tracking a given portfolio. The function checks if the input is a valid 
`Portfolio` object and filters SMAs based on their base portfolio's short name.
}
\details{
The function first validates that the input is of class `Portfolio`. 
  It then retrieves all portfolios from the `registries$portfolios` environment 
  and filters those that are of class `SMA`. Finally, it checks which SMAs 
  have the given portfolio as their base portfolio and returns them.
}
