% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-loading.R
\name{load_all_portfolios_from_db}
\alias{load_all_portfolios_from_db}
\title{Load All Portfolios and SMAs from Database}
\usage{
load_all_portfolios_from_db(con = get_db_connection())
}
\arguments{
\item{con}{DBI connection}
}
\value{
Named list of all Portfolio/SMA objects (invisibly)
}
\description{
Load All Portfolios and SMAs from Database
}
