% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-overflowrule.R
\name{OverflowRule}
\alias{OverflowRule}
\title{OverflowRule}
\description{
OverflowRule

OverflowRule
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OverflowRule-new}{\code{OverflowRule$new()}}
\item \href{#method-OverflowRule-build_constraints}{\code{OverflowRule$build_constraints()}}
\item \href{#method-OverflowRule-objective_terms}{\code{OverflowRule$objective_terms()}}
\item \href{#method-OverflowRule-clone}{\code{OverflowRule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OverflowRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-OverflowRule-new}{}}}
\subsection{Method \code{new()}}{
Create a new OverflowRule R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverflowRule$new(replacements)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replacements}}{Named list of replacements. Each name is a source
security ID, and each element is a character vector of target security
IDs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OverflowRule-build_constraints"></a>}}
\if{latex}{\out{\hypertarget{method-OverflowRule-build_constraints}{}}}
\subsection{Method \code{build_constraints()}}{
Build CVXR constraints for the rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverflowRule$build_constraints(ctx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ctx}}{Context object with optimization variables and parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OverflowRule-objective_terms"></a>}}
\if{latex}{\out{\hypertarget{method-OverflowRule-objective_terms}{}}}
\subsection{Method \code{objective_terms()}}{
Objective terms contributed by this rule (Dummy)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverflowRule$objective_terms(ctx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ctx}}{Context object with optimization variables and parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An empty list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OverflowRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OverflowRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverflowRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
