% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio (R6 Object)}
\description{
R6 Class representing a portfolio object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Portfolio Id}

\item{\code{long_name}}{Portfolio Long Name}

\item{\code{short_name}}{Portfolio Short Name}

\item{\code{nav}}{Portfolio NAV}

\item{\code{enfusion_url}}{Enfusion Web URL to
Consolidated Position Listing Report}

\item{\code{positions}}{List of Positions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Portfolio-new}{\code{Portfolio$new()}}
\item \href{#method-Portfolio-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-new"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-new}{}}}
\subsection{Method \code{new()}}{
Create New PortfolioR6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(long_name, short_name, enfusion_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Portfolio Long Name}

\item{\code{short_name}}{Portfolio Short Name}

\item{\code{enfusion_url}}{Enfusion URL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
