% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio (R6 Object)}
\description{
R6 Class representing a portfolio object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Portfolio-new}{\code{Portfolio$new()}}
\item \href{#method-Portfolio-get_short_name}{\code{Portfolio$get_short_name()}}
\item \href{#method-Portfolio-get_nav}{\code{Portfolio$get_nav()}}
\item \href{#method-Portfolio-get_position}{\code{Portfolio$get_position()}}
\item \href{#method-Portfolio-get_target_position}{\code{Portfolio$get_target_position()}}
\item \href{#method-Portfolio-get_rules}{\code{Portfolio$get_rules()}}
\item \href{#method-Portfolio-get_trade_constructor}{\code{Portfolio$get_trade_constructor()}}
\item \href{#method-Portfolio-get_replacement_security}{\code{Portfolio$get_replacement_security()}}
\item \href{#method-Portfolio-get_replaced_security}{\code{Portfolio$get_replaced_security()}}
\item \href{#method-Portfolio-check_rules_target}{\code{Portfolio$check_rules_target()}}
\item \href{#method-Portfolio-check_rules_current}{\code{Portfolio$check_rules_current()}}
\item \href{#method-Portfolio-get_security_position_limits}{\code{Portfolio$get_security_position_limits()}}
\item \href{#method-Portfolio-get_swap_flag_position_rules}{\code{Portfolio$get_swap_flag_position_rules()}}
\item \href{#method-Portfolio-calc_proposed_trade}{\code{Portfolio$calc_proposed_trade()}}
\item \href{#method-Portfolio-add_flow}{\code{Portfolio$add_flow()}}
\item \href{#method-Portfolio-add_position}{\code{Portfolio$add_position()}}
\item \href{#method-Portfolio-add_target_position}{\code{Portfolio$add_target_position()}}
\item \href{#method-Portfolio-remove_target_position}{\code{Portfolio$remove_target_position()}}
\item \href{#method-Portfolio-add_rule}{\code{Portfolio$add_rule()}}
\item \href{#method-Portfolio-add_replacement}{\code{Portfolio$add_replacement()}}
\item \href{#method-Portfolio-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-new"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-new}{}}}
\subsection{Method \code{new()}}{
Create New Portfolio R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(long_name, short_name, nav, positions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Portfolio Long Name}

\item{\code{short_name}}{Portfolio Short Name}

\item{\code{nav}}{NAV of portfolio}

\item{\code{positions}}{list of position items}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_short_name"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_short_name}{}}}
\subsection{Method \code{get_short_name()}}{
Get Portfolio short name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_short_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_nav"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_nav}{}}}
\subsection{Method \code{get_nav()}}{
Get Fund NAV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_nav()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_position}{}}}
\subsection{Method \code{get_position()}}{
Get list of positions in portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_position(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Ticker}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_target_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_target_position}{}}}
\subsection{Method \code{get_target_position()}}{
Get list of target positions in portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_target_position(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Ticker}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_rules}{}}}
\subsection{Method \code{get_rules()}}{
Get the Rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_rules()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of rules
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_trade_constructor"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_trade_constructor}{}}}
\subsection{Method \code{get_trade_constructor()}}{
Get the Trade Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_trade_constructor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The portfolio constructor object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_replacement_security"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_replacement_security}{}}}
\subsection{Method \code{get_replacement_security()}}{
Get replacement security for a given replaced security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_replacement_security(replaced_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replaced_security_id}}{Security ID of the replaced security (in base ptfl) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_replaced_security"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_replaced_security}{}}}
\subsection{Method \code{get_replaced_security()}}{
Get replaced security for a given replacement security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_replaced_security(replacement_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replacement_security_id}}{Security ID of the replacement security (in SMA) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-check_rules_target"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-check_rules_target}{}}}
\subsection{Method \code{check_rules_target()}}{
Check SMA rules against the target positions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$check_rules_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-check_rules_current"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-check_rules_current}{}}}
\subsection{Method \code{check_rules_current()}}{
Check SMA rules against the current positions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$check_rules_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_security_position_limits"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_security_position_limits}{}}}
\subsection{Method \code{get_security_position_limits()}}{
Get Max and Min Value of the security given all SMA Rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_security_position_limits(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_swap_flag_position_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_swap_flag_position_rules}{}}}
\subsection{Method \code{get_swap_flag_position_rules()}}{
Get Swap Flag for a given security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_swap_flag_position_rules(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-calc_proposed_trade"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-calc_proposed_trade}{}}}
\subsection{Method \code{calc_proposed_trade()}}{
Calculate the trade quantity for a given security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$calc_proposed_trade(security_id = NULL, trade_qty = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}

\item{\code{trade_qty}}{Trade quantity (default: 0)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_flow"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_flow}{}}}
\subsection{Method \code{add_flow()}}{
Add flow to portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_flow(flow = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{flow amount}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_position}{}}}
\subsection{Method \code{add_position()}}{
Add Position to Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_position(position, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{Position S6 Object}

\item{\code{overwrite}}{Logical. Overwrite existing position if TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_target_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_target_position}{}}}
\subsection{Method \code{add_target_position()}}{
Add New Target Position to Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_target_position(position, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{Position S6 Object}

\item{\code{overwrite}}{Logical. Overwrite existing position if TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-remove_target_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-remove_target_position}{}}}
\subsection{Method \code{remove_target_position()}}{
Remove existing Target Position from Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$remove_target_position(position_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position_id}}{Position ID
Add Rule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_rule"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_rule}{}}}
\subsection{Method \code{add_rule()}}{
Create Rule and Add to Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_rule(rule)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rule}}{An object of class SMARule
Add Replacement}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_replacement"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_replacement}{}}}
\subsection{Method \code{add_replacement()}}{
Add replacement securitity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_replacement(
  original_security = NULL,
  replacement_security = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{original_security}}{The original Security id}

\item{\code{replacement_security}}{The replacement Security id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
