% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio (R6 Object)}
\description{
R6 Class representing a portfolio object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Portfolio-new}{\code{Portfolio$new()}}
\item \href{#method-Portfolio-get_short_name}{\code{Portfolio$get_short_name()}}
\item \href{#method-Portfolio-get_nav}{\code{Portfolio$get_nav()}}
\item \href{#method-Portfolio-get_position}{\code{Portfolio$get_position()}}
\item \href{#method-Portfolio-get_rules}{\code{Portfolio$get_rules()}}
\item \href{#method-Portfolio-get_trade_constructor}{\code{Portfolio$get_trade_constructor()}}
\item \href{#method-Portfolio-get_replacement_security}{\code{Portfolio$get_replacement_security()}}
\item \href{#method-Portfolio-get_replaced_security}{\code{Portfolio$get_replaced_security()}}
\item \href{#method-Portfolio-get_security_position_limits}{\code{Portfolio$get_security_position_limits()}}
\item \href{#method-Portfolio-get_swap_flag_position_rules}{\code{Portfolio$get_swap_flag_position_rules()}}
\item \href{#method-Portfolio-add_flow}{\code{Portfolio$add_flow()}}
\item \href{#method-Portfolio-add_position}{\code{Portfolio$add_position()}}
\item \href{#method-Portfolio-add_rule}{\code{Portfolio$add_rule()}}
\item \href{#method-Portfolio-add_replacement}{\code{Portfolio$add_replacement()}}
\item \href{#method-Portfolio-update_enfusion}{\code{Portfolio$update_enfusion()}}
\item \href{#method-Portfolio-rebalance}{\code{Portfolio$rebalance()}}
\item \href{#method-Portfolio-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-new"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-new}{}}}
\subsection{Method \code{new()}}{
Create New Portfolio R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(long_name, short_name, holdings_url, nav, positions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Portfolio Long Name}

\item{\code{short_name}}{Portfolio Short Name}

\item{\code{holdings_url}}{URL to Enfusion Holdings Report}

\item{\code{nav}}{NAV of portfolio}

\item{\code{positions}}{list of position items}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_short_name"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_short_name}{}}}
\subsection{Method \code{get_short_name()}}{
Get Portfolio short name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_short_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_nav"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_nav}{}}}
\subsection{Method \code{get_nav()}}{
Get Fund NAV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_nav()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_position}{}}}
\subsection{Method \code{get_position()}}{
Get list of positions in portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_position(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Ticker}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_rules}{}}}
\subsection{Method \code{get_rules()}}{
Get the Rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_rules()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of rules
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_trade_constructor"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_trade_constructor}{}}}
\subsection{Method \code{get_trade_constructor()}}{
Get the Trade Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_trade_constructor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The portfolio constructor object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_replacement_security"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_replacement_security}{}}}
\subsection{Method \code{get_replacement_security()}}{
Get replacement security for a given replaced security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_replacement_security(replaced_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replaced_security_id}}{Security ID of the replaced security (in base ptfl) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_replaced_security"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_replaced_security}{}}}
\subsection{Method \code{get_replaced_security()}}{
Get replaced security for a given replacement security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_replaced_security(replacement_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replacement_security_id}}{Security ID of the replacement security (in SMA) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_security_position_limits"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_security_position_limits}{}}}
\subsection{Method \code{get_security_position_limits()}}{
Get Max and Min Value of the security given all SMA Rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_security_position_limits(
  security_id = NULL,
  update_bbfields = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}

\item{\code{update_bbfields}}{Logical. Update Bloomberg fields (default: TRUE) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-get_swap_flag_position_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-get_swap_flag_position_rules}{}}}
\subsection{Method \code{get_swap_flag_position_rules()}}{
Get Swap Flag for a given security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$get_swap_flag_position_rules(
  security_id = NULL,
  update_bbfields = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}

\item{\code{update_bbfields}}{Logical. Update Bloomberg fields (default: TRUE) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_flow"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_flow}{}}}
\subsection{Method \code{add_flow()}}{
Add flow to portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_flow(flow = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{flow amount}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_position"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_position}{}}}
\subsection{Method \code{add_position()}}{
Add Position to Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_position(position, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{Position S6 Object}

\item{\code{overwrite}}{Logical. Overwrite existing position if TRUE
Add Rule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_rule"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_rule}{}}}
\subsection{Method \code{add_rule()}}{
Create Rule and Add to Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_rule(rule)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rule}}{An object of class SMARule
Add Replacement}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-add_replacement"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-add_replacement}{}}}
\subsection{Method \code{add_replacement()}}{
Add replacement securitity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$add_replacement(
  original_security = NULL,
  replacement_security = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{original_security}}{The original Security id}

\item{\code{replacement_security}}{The replacement Security id
Update Enfusion Data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-update_enfusion"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-update_enfusion}{}}}
\subsection{Method \code{update_enfusion()}}{
Update Positions
Rebalance Portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$update_enfusion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-rebalance"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-rebalance}{}}}
\subsection{Method \code{rebalance()}}{
Calculate the trade quantity for a given security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$rebalance(update_bbfields = TRUE, as.df = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{update_bbfields}}{Logical. Update Bloomberg fields (default: TRUE)}

\item{\code{as.df}}{Logical. Return as data frame (default: TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
