% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-position.R
\name{Position}
\alias{Position}
\title{Position (R6 Object)}
\description{
A position
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ticker}}{Security Ticker}

\item{\code{desc}}{Security Description}

\item{\code{stock_qty}}{Stock Quantity}

\item{\code{delta_qty}}{Delta Quantity}

\item{\code{total_qty}}{Total Quantity}

\item{\code{mkt_val}}{Market Value}

\item{\code{delta_val}}{Delta Value}

\item{\code{stock_pct_nav}}{Stock Percent of NAV}

\item{\code{delta_pct_nav}}{Delta PErcent of NAV}

\item{\code{security}}{SecurityR6 Objct}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PoisitionR6-new}{\code{Position$new()}}
\item \href{#method-PoisitionR6-clone}{\code{Position$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoisitionR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-PoisitionR6-new}{}}}
\subsection{Method \code{new()}}{
Create New PositionR6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$new(
  ticker,
  desc,
  stock_qty,
  delta_qty,
  total_qty,
  mkt_val,
  delta_val,
  stock_pct_nav,
  delta_pct_nav
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Security Ticker}

\item{\code{desc}}{Security Description}

\item{\code{stock_qty}}{Stock Quantity}

\item{\code{delta_qty}}{Delta Quantity}

\item{\code{total_qty}}{Total Quantity}

\item{\code{mkt_val}}{Market Value}

\item{\code{delta_val}}{Delta Value}

\item{\code{stock_pct_nav}}{Stock Percent of NAV}

\item{\code{delta_pct_nav}}{Delta PErcent of NAV}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PoisitionR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PoisitionR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
