% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-security.R
\docType{class}
\name{Security}
\alias{Security}
\title{Security (R6 Class)}
\description{
R6 class for a security with time-series and non-time-series data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Security-new}{\code{Security$new()}}
\item \href{#method-Security-set_data_item}{\code{Security$set_data_item()}}
\item \href{#method-Security-get_data_item}{\code{Security$get_data_item()}}
\item \href{#method-Security-get_ts_data_item}{\code{Security$get_ts_data_item()}}
\item \href{#method-Security-clone}{\code{Security$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-new"></a>}}
\if{latex}{\out{\hypertarget{method-Security-new}{}}}
\subsection{Method \code{new()}}{
Create new Security R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$new(id = NULL, ts_data = NULL, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Character string. Security identifier (e.g., "AAPL").}

\item{\code{ts_data}}{Named list of time-series data (like volume, prices).}

\item{\code{data}}{Named list of non-time-series data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-set_data_item"></a>}}
\if{latex}{\out{\hypertarget{method-Security-set_data_item}{}}}
\subsection{Method \code{set_data_item()}}{
Add or overwrite an item in \code{data}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$set_data_item(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named values to store in \code{data} list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-get_data_item"></a>}}
\if{latex}{\out{\hypertarget{method-Security-get_data_item}{}}}
\subsection{Method \code{get_data_item()}}{
Get an items from \code{data} as a named value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$get_data_item(item = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{String name of the list item in \code{data}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-get_ts_data_item"></a>}}
\if{latex}{\out{\hypertarget{method-Security-get_ts_data_item}{}}}
\subsection{Method \code{get_ts_data_item()}}{
Get an item from \code{ts_data} as a named vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$get_ts_data_item(item = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{String name of the time-series vector in \code{ts_data}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Security-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
