% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-security.R
\docType{class}
\name{Security}
\alias{Security}
\title{Security (R6 Object)}
\description{
A security where data comes from yahoo finance
}
\examples{

## ------------------------------------------------
## Method `Security$get_adv`
## ------------------------------------------------

\dontrun{
  sec <- Security$new("AAPL")
  sec$get_adv(30)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ticker}}{Security Ticker}

\item{\code{yahoo_match}}{Boolean if ticker was found on yahoo finance}

\item{\code{long_name}}{Security Long Name}

\item{\code{short_name}}{Security Short Name}

\item{\code{ts_data}}{Time Series Data List}

\item{\code{data}}{Non-Time Series Data List}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Security-new}{\code{Security$new()}}
\item \href{#method-Security-get_ts_data_item}{\code{Security$get_ts_data_item()}}
\item \href{#method-Security-get_adv}{\code{Security$get_adv()}}
\item \href{#method-Security-clone}{\code{Security$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-new"></a>}}
\if{latex}{\out{\hypertarget{method-Security-new}{}}}
\subsection{Method \code{new()}}{
Create New Security R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$new(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Security Ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Security} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-get_ts_data_item"></a>}}
\if{latex}{\out{\hypertarget{method-Security-get_ts_data_item}{}}}
\subsection{Method \code{get_ts_data_item()}}{
Gets an item from the ts_data list field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$get_ts_data_item(item = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{String. Item in ts_data list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An array of ts_data with dates as name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-get_adv"></a>}}
\if{latex}{\out{\hypertarget{method-Security-get_adv}{}}}
\subsection{Method \code{get_adv()}}{
Computes the average daily volume for the specified number
 of trailing days.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$get_adv(adv_days = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adv_days}}{Integer. The number of days to consider.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method is added onto the \code{Security} R6 class dynamically.
}

\subsection{Returns}{
A numeric value representing the average daily volume.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  sec <- Security$new("AAPL")
  sec$get_adv(30)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Security-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Security-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Security$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
