% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sma.R
\name{SMA}
\alias{SMA}
\title{SMA (R6 Object)}
\description{
R6 Class representing a seperately managed account.
 A seperately managed account is linked to a portfolio and contains
 Sma_Rules.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{SMA Id}

\item{\code{long_name}}{SMA Long Name}

\item{\code{short_name}}{SMA Short Name}

\item{\code{nav}}{SMA NAV}

\item{\code{enfusion_url}}{Enfusion Web URL to
Consolidated Position Listing Report}

\item{\code{positions}}{List of Positions}

\item{\code{target_portfolio}}{Target Portfolio}

\item{\code{sma_rules}}{SMA Rules}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMA-new}{\code{SMA$new()}}
\item \href{#method-SMA-clone}{\code{SMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$new(long_name, short_name, enfusion_url, target_portfolio)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
