% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sma.R
\name{SMA}
\alias{SMA}
\title{SMA (R6 Object)}
\description{
R6 Class representing a seperately managed account.
 A seperately managed account is linked to a portfolio and contains
 Rules.
}
\section{Super class}{
\code{\link[SMAManager:Portfolio]{SMAManager::Portfolio}} -> \code{SMA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMA-new}{\code{SMA$new()}}
\item \href{#method-SMA-get_base_portfolio}{\code{SMA$get_base_portfolio()}}
\item \href{#method-SMA-get_base_portfolio_position}{\code{SMA$get_base_portfolio_position()}}
\item \href{#method-SMA-check_rule_compliance}{\code{SMA$check_rule_compliance()}}
\item \href{#method-SMA-clone}{\code{SMA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_flow"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_flow'><code>SMAManager::Portfolio$add_flow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_position'><code>SMAManager::Portfolio$add_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_replacement"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_replacement'><code>SMAManager::Portfolio$add_replacement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_rule"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_rule'><code>SMAManager::Portfolio$add_rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_nav"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_nav'><code>SMAManager::Portfolio$get_nav()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_position'><code>SMAManager::Portfolio$get_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_replaced_security"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_replaced_security'><code>SMAManager::Portfolio$get_replaced_security()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_replacement_security"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_replacement_security'><code>SMAManager::Portfolio$get_replacement_security()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_rules"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_rules'><code>SMAManager::Portfolio$get_rules()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_security_position_limits"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_security_position_limits'><code>SMAManager::Portfolio$get_security_position_limits()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_short_name"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_short_name'><code>SMAManager::Portfolio$get_short_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_swap_flag_position_rules"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_swap_flag_position_rules'><code>SMAManager::Portfolio$get_swap_flag_position_rules()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_trade_constructor"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_trade_constructor'><code>SMAManager::Portfolio$get_trade_constructor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="rebalance"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-rebalance'><code>SMAManager::Portfolio$rebalance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="update_enfusion"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-update_enfusion'><code>SMAManager::Portfolio$update_enfusion()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-new}{}}}
\subsection{Method \code{new()}}{
Create a new SMA R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$new(
  long_name,
  short_name,
  holdings_url,
  nav,
  positions = NULL,
  base_portfolio = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Character. SMA Long Name.}

\item{\code{short_name}}{Character. SMA Short Name.}

\item{\code{holdings_url}}{Character. URL to Enfusion Holdings Report.}

\item{\code{nav}}{Numeric. SMA Net Asset Value.}

\item{\code{positions}}{Optional. SMA Positions. Default is NULL.}

\item{\code{base_portfolio}}{An object representing the base portfolio.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the SMA class.
Get Base Portfolio
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_base_portfolio"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_base_portfolio}{}}}
\subsection{Method \code{get_base_portfolio()}}{
Get the tagret portfolio
Get Base Portfolio Position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_base_portfolio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_base_portfolio_position"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_base_portfolio_position}{}}}
\subsection{Method \code{get_base_portfolio_position()}}{
Get a position in the Base portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_base_portfolio_position(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID
Check Rule Compliance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-check_rule_compliance"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-check_rule_compliance}{}}}
\subsection{Method \code{check_rule_compliance()}}{
Check if the SMA is compliant with all its rules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$check_rule_compliance(update_bbfields = TRUE, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{update_bbfields}}{Logical. Whether to update Bloomberg data before
checking rules. Defaults to TRUE.}

\item{\code{verbose}}{Logical. Whether to print compliance results. Defaults to 
FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of rule compliance results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
