% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sma.R
\name{SMA}
\alias{SMA}
\title{SMA (R6 Object)}
\description{
R6 Class representing a seperately managed account.
 A seperately managed account is linked to a portfolio and contains
 Sma_Rules.
}
\section{Super class}{
\code{\link[SMAManager:Portfolio]{SMAManager::Portfolio}} -> \code{SMA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMA-new}{\code{SMA$new()}}
\item \href{#method-SMA-add_rule}{\code{SMA$add_rule()}}
\item \href{#method-SMA-add_replacement}{\code{SMA$add_replacement()}}
\item \href{#method-SMA-get_target_portfolio}{\code{SMA$get_target_portfolio()}}
\item \href{#method-SMA-get_target_portfolio_position}{\code{SMA$get_target_portfolio_position()}}
\item \href{#method-SMA-compare_weight_current}{\code{SMA$compare_weight_current()}}
\item \href{#method-SMA-get_sma_rules}{\code{SMA$get_sma_rules()}}
\item \href{#method-SMA-get_replacement_security}{\code{SMA$get_replacement_security()}}
\item \href{#method-SMA-get_replaced_security}{\code{SMA$get_replaced_security()}}
\item \href{#method-SMA-clone}{\code{SMA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_target_position'><code>SMAManager::Portfolio$add_target_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_nav"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_nav'><code>SMAManager::Portfolio$get_nav()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_position'><code>SMAManager::Portfolio$get_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_short_name"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_short_name'><code>SMAManager::Portfolio$get_short_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_target_position'><code>SMAManager::Portfolio$get_target_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="print"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-print'><code>SMAManager::Portfolio$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="remove_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-remove_target_position'><code>SMAManager::Portfolio$remove_target_position()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-new}{}}}
\subsection{Method \code{new()}}{
Create a new SMA R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$new(long_name, short_name, nav, target_portfolio = NULL, positions = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Character. SMA Long Name.}

\item{\code{short_name}}{Character. SMA Short Name.}

\item{\code{nav}}{Numeric. SMA Net Asset Value.}

\item{\code{target_portfolio}}{An object representing the target portfolio.}

\item{\code{positions}}{Optional. SMA Positions. Default is NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the SMA class.
Add SMA Rule
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-add_rule"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-add_rule}{}}}
\subsection{Method \code{add_rule()}}{
Create SMA Rule and Add to SMA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$add_rule(rule_name, rule_scope, rule_formula)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rule_name}}{Character representing identifing name of rule}

\item{\code{rule_scope}}{Character representing the scope of the rule}

\item{\code{rule_formula}}{formula representing the rule
Add Replacement}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-add_replacement"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-add_replacement}{}}}
\subsection{Method \code{add_replacement()}}{
Add replacement securitity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$add_replacement(original_security = NULL, replacement_security = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{original_security}}{The original Security id}

\item{\code{replacement_security}}{The replacement Security id
Get Target Portfolio}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_target_portfolio"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_target_portfolio}{}}}
\subsection{Method \code{get_target_portfolio()}}{
Get the tagret portfolio
Get Target Portfolio Position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_target_portfolio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_target_portfolio_position"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_target_portfolio_position}{}}}
\subsection{Method \code{get_target_portfolio_position()}}{
Get a position in the target portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_target_portfolio_position(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID
Compare Weights to Target Portfolio}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-compare_weight_current"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-compare_weight_current}{}}}
\subsection{Method \code{compare_weight_current()}}{
Compare the current weight in the sma
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$compare_weight_current(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID
to the current weight in the target portfolio}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_sma_rules"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_sma_rules}{}}}
\subsection{Method \code{get_sma_rules()}}{
Get the SMA Rules
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_sma_rules()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of SMA rules
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_replacement_security"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_replacement_security}{}}}
\subsection{Method \code{get_replacement_security()}}{
Get replacement security for a given replaced security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_replacement_security(replaced_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replaced_security_id}}{Security ID of the replaced security (in target ptfl) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_replaced_security"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_replaced_security}{}}}
\subsection{Method \code{get_replaced_security()}}{
Get replaced security for a given replacement security
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_replaced_security(replacement_security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replacement_security_id}}{Security ID of the replacement security (in SMA) #nolint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
