% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sma.R
\name{SMA}
\alias{SMA}
\title{SMA (R6 Object)}
\description{
R6 Class representing a seperately managed account.
 A seperately managed account is linked to a portfolio and contains
 Rules.
}
\section{Super class}{
\code{\link[SMAManager:Portfolio]{SMAManager::Portfolio}} -> \code{SMA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMA-new}{\code{SMA$new()}}
\item \href{#method-SMA-get_base_portfolio}{\code{SMA$get_base_portfolio()}}
\item \href{#method-SMA-get_base_portfolio_position}{\code{SMA$get_base_portfolio_position()}}
\item \href{#method-SMA-mimic_base_portfolio}{\code{SMA$mimic_base_portfolio()}}
\item \href{#method-SMA-clone}{\code{SMA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_flow"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_flow'><code>SMAManager::Portfolio$add_flow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_position'><code>SMAManager::Portfolio$add_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_replacement"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_replacement'><code>SMAManager::Portfolio$add_replacement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_rule"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_rule'><code>SMAManager::Portfolio$add_rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="add_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-add_target_position'><code>SMAManager::Portfolio$add_target_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="calc_proposed_trade"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-calc_proposed_trade'><code>SMAManager::Portfolio$calc_proposed_trade()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="check_rules_target"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-check_rules_target'><code>SMAManager::Portfolio$check_rules_target()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_nav"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_nav'><code>SMAManager::Portfolio$get_nav()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_position'><code>SMAManager::Portfolio$get_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_replaced_security"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_replaced_security'><code>SMAManager::Portfolio$get_replaced_security()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_replacement_security"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_replacement_security'><code>SMAManager::Portfolio$get_replacement_security()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_rules"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_rules'><code>SMAManager::Portfolio$get_rules()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_security_position_limits"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_security_position_limits'><code>SMAManager::Portfolio$get_security_position_limits()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_short_name"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_short_name'><code>SMAManager::Portfolio$get_short_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_swap_flag_position_rules"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_swap_flag_position_rules'><code>SMAManager::Portfolio$get_swap_flag_position_rules()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_target_position'><code>SMAManager::Portfolio$get_target_position()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="get_trade_constructor"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-get_trade_constructor'><code>SMAManager::Portfolio$get_trade_constructor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="Portfolio" data-id="remove_target_position"><a href='../../SMAManager/html/Portfolio.html#method-Portfolio-remove_target_position'><code>SMAManager::Portfolio$remove_target_position()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-new}{}}}
\subsection{Method \code{new()}}{
Create a new SMA R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$new(long_name, short_name, nav, positions = NULL, base_portfolio = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_name}}{Character. SMA Long Name.}

\item{\code{short_name}}{Character. SMA Short Name.}

\item{\code{nav}}{Numeric. SMA Net Asset Value.}

\item{\code{positions}}{Optional. SMA Positions. Default is NULL.}

\item{\code{base_portfolio}}{An object representing the base portfolio.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the SMA class.
Get Base Portfolio
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_base_portfolio"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_base_portfolio}{}}}
\subsection{Method \code{get_base_portfolio()}}{
Get the tagret portfolio
Get Base Portfolio Position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_base_portfolio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-get_base_portfolio_position"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-get_base_portfolio_position}{}}}
\subsection{Method \code{get_base_portfolio_position()}}{
Get a position in the Base portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$get_base_portfolio_position(security_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-mimic_base_portfolio"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-mimic_base_portfolio}{}}}
\subsection{Method \code{mimic_base_portfolio()}}{
mimic the base portfolio target position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$mimic_base_portfolio(security_id = NULL, assign_to_registry = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}

\item{\code{assign_to_registry}}{Assign to registry (default: TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
