% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-smarule.R
\name{SMARule}
\alias{SMARule}
\title{SMA Rule (R6 Onject)}
\description{
R6 Class that encapsultes rules for SMAs.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMARule-new}{\code{SMARule$new()}}
\item \href{#method-SMARule-get_id}{\code{SMARule$get_id()}}
\item \href{#method-SMARule-get_name}{\code{SMARule$get_name()}}
\item \href{#method-SMARule-get_scope}{\code{SMARule$get_scope()}}
\item \href{#method-SMARule-get_definition}{\code{SMARule$get_definition()}}
\item \href{#method-SMARule-get_threshold}{\code{SMARule$get_threshold()}}
\item \href{#method-SMARule-clone}{\code{SMARule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$new(
  sma_name = NULL,
  name = NULL,
  scope = NULL,
  definition = NULL,
  threshold = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sma_name}}{Character}

\item{\code{name}}{Character}

\item{\code{scope}}{Character}

\item{\code{definition}}{Formula}

\item{\code{threshold}}{Formula
Get Id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get Id of SMA Rule
Get Name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$get_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get name of SMA Rule
Get Scope
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-get_scope"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-get_scope}{}}}
\subsection{Method \code{get_scope()}}{
Get the scope of the SMA Rule
Get Definition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$get_scope()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-get_definition"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-get_definition}{}}}
\subsection{Method \code{get_definition()}}{
Get the definition of the SMA Rule
Get Threshold
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$get_definition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-get_threshold"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-get_threshold}{}}}
\subsection{Method \code{get_threshold()}}{
Get the threshold of the SMA Rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$get_threshold()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMARule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
