% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-smaruleportfolio.R
\name{SMARulePortfolio}
\alias{SMARulePortfolio}
\title{SMA Rule for Portfolio}
\description{
SMA Rule for Portfolio

SMA Rule for Portfolio
}
\section{Super class}{
\code{\link[SMAManager:SMARule]{SMAManager::SMARule}} -> \code{SMARulePortfolio}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMARulePortfolio-check_rule_current}{\code{SMARulePortfolio$check_rule_current()}}
\item \href{#method-SMARulePortfolio-check_rule_target}{\code{SMARulePortfolio$check_rule_target()}}
\item \href{#method-SMARulePortfolio-check_swap_current}{\code{SMARulePortfolio$check_swap_current()}}
\item \href{#method-SMARulePortfolio-check_swap_target}{\code{SMARulePortfolio$check_swap_target()}}
\item \href{#method-SMARulePortfolio-get_security_limits}{\code{SMARulePortfolio$get_security_limits()}}
\item \href{#method-SMARulePortfolio-clone}{\code{SMARulePortfolio$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_definition"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_definition'><code>SMAManager::SMARule$get_definition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_id"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_id'><code>SMAManager::SMARule$get_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_max_threshold"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_max_threshold'><code>SMAManager::SMARule$get_max_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_min_threshold"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_min_threshold'><code>SMAManager::SMARule$get_min_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_name"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_name'><code>SMAManager::SMARule$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_scope"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_scope'><code>SMAManager::SMARule$get_scope()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_swap_only"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_swap_only'><code>SMAManager::SMARule$get_swap_only()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="initialize"><a href='../../SMAManager/html/SMARule.html#method-SMARule-initialize'><code>SMAManager::SMARule$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-check_rule_current"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-check_rule_current}{}}}
\subsection{Method \code{check_rule_current()}}{
Check the rule against the current portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$check_rule_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-check_rule_target"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-check_rule_target}{}}}
\subsection{Method \code{check_rule_target()}}{
Check the rule against the target portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$check_rule_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-check_swap_current"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-check_swap_current}{}}}
\subsection{Method \code{check_swap_current()}}{
Check the swap rule against the current portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$check_swap_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-check_swap_target"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-check_swap_target}{}}}
\subsection{Method \code{check_swap_target()}}{
Check the swap rule against the target portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$check_swap_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-get_security_limits"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-get_security_limits}{}}}
\subsection{Method \code{get_security_limits()}}{
Get the Max and Min value of the security based on the rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$get_security_limits(security_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{security_id}}{Security ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of Max and Min Value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePortfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePortfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePortfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
