% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-smaruleposition.R
\name{SMARulePosition}
\alias{SMARulePosition}
\title{SMA Rule for Positions}
\description{
R6 Class that encapsulates rule for SMAs
}
\section{Super class}{
\code{\link[SMAManager:SMARule]{SMAManager::SMARule}} -> \code{SMARulePosition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMARulePosition-get_sma}{\code{SMARulePosition$get_sma()}}
\item \href{#method-SMARulePosition-check_rule_current}{\code{SMARulePosition$check_rule_current()}}
\item \href{#method-SMARulePosition-clone}{\code{SMARulePosition$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_definition"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_definition'><code>SMAManager::SMARule$get_definition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_id"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_id'><code>SMAManager::SMARule$get_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_name"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_name'><code>SMAManager::SMARule$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_scope"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_scope'><code>SMAManager::SMARule$get_scope()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="get_threshold"><a href='../../SMAManager/html/SMARule.html#method-SMARule-get_threshold'><code>SMAManager::SMARule$get_threshold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SMAManager" data-topic="SMARule" data-id="initialize"><a href='../../SMAManager/html/SMARule.html#method-SMARule-initialize'><code>SMAManager::SMARule$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePosition-get_sma"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePosition-get_sma}{}}}
\subsection{Method \code{get_sma()}}{
Get SMA the Rule applies to
Check Rule Against Current Holdings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePosition$get_sma()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePosition-check_rule_current"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePosition-check_rule_current}{}}}
\subsection{Method \code{check_rule_current()}}{
Check the rule against the current holdings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePosition$check_rule_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMARulePosition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMARulePosition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMARulePosition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
