% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_bloomberg_fields}
\alias{update_bloomberg_fields}
\title{Add Bloomberg Data for SMA Rules to Securities}
\usage{
update_bloomberg_fields()
}
\value{
TRUE (invisible)
}
\description{
Add Bloomberg Data for SMA Rules to Securities
}
